refactor:move user repository
