+++ snowtricks/src/Auth/Domain/Repository/UserRepository.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Auth;
+namespace App\Auth\Domain\Repository;
 
 use App\Auth\Domain\Entity\User;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
